Take only the kind each + asks for, and add a 62.5% rung
The collection + accepted an MP3 and parsed 13 MB of audio as M3U: 52,454
garbage tracks, 27 MB of playlist_tracks.json, and a restart that painted
those bytes as a saved playlist. Three separate things had to be true for
that, and all three are fixed.

The Linux portal is the default picker and never sent our filters at all --
portalPick built its options vardict without them -- so the audio-only and
playlist-only filters both platforms honour were dead on the platform most
listeners use. They go out now as a(sa(us)), with each letter as a class
because portal globs are case-sensitive and Track.MP3 would otherwise be
hidden by its own filter. The extension lists stay in files.cpp, which is
what audioExtensions() and playlistExtensions() were for; session.cpp had
been re-typing them as literals.

A filter is advice, and the portal spec says so in as many words: it may
hand back anything. So the two pickers re-check what they asked for and
treat a mismatch as a cancel. That is what makes "playlists only from the
PLAYLISTS column" true rather than merely displayed.

Ingest refuses content that is not playlist text. A NUL byte is the tell --
no filesystem allows one in a name -- so an audio file named .m3u no longer
becomes thousands of rows of itself, and the ~171,000 QFileInfo warnings it
provoked go with it. Refresh leaves the current list standing on such a file
instead of emptying it, since that is no answer, and it decides that before
raising the discard prompt rather than after. M3uCodec::resolve walks at most
32 trailing segments: its tail search is quadratic in the slashes on a line,
which is 32 s for a 64 KB line and about 2.3 hours for 2 MB. A cached track
path carrying a NUL is dropped on read, so a collection already poisoned by
an older build stops replaying it.

62.5% sits between 50 and 75 because one press used to cut the panels by a
third with nothing in between. The percent was an int everywhere, so the
ladder is qreal now: zoomPercentsEqual recognises a rung rather than ==,
zoomLabel is the one spelling the readout, the tips and the bench lines
share, and settings.json reads through toDouble -- toInt refuses a
fractional JSON number and would have dropped 62.5 onto the default. A whole
number written by an older build still loads and still snaps, and 62 now
snaps up to 62.5 rather than down to 50. It is also the first step whose
height is fractional (217.5), and the only one whose readout is five
characters, so --dump-chrome photographs it against the slot the title
cluster reserves.

Double-clicking a saved playlist renames it. A single click loads, so the
load waits out the platform double-click interval and the second click
cancels it -- the one place in the chrome where a press is not dispatched
at once. The dialog says the file on disk keeps its name, because a listener
who suspects otherwise will not use the feature.

+++ src/chrome_menu.cpp
@@ -52,7 +52,7 @@ void drawCheck(QPainter& p, QPointF centre, const QColor& colour, qreal zoom) {
 
 class ChromeMenuWindow : public QWidget {
  public:
-  ChromeMenuWindow(QWidget* owner, const QVector<ChromeMenuItem>& items, int zoomPercent,
+  ChromeMenuWindow(QWidget* owner, const QVector<ChromeMenuItem>& items, qreal zoomPercent,
                    const ChromeTokens& look)
       // Qt::Popup is the only window type Qt can grab the pointer for on
       // Wayland ("This plugin supports grabbing the mouse only for popup
@@ -72,7 +72,7 @@ class ChromeMenuWindow : public QWidget {
       // compositor ends up hanging this off is the shell, which is what we want.
       : QWidget(owner, Qt::Popup | Qt::FramelessWindowHint | Qt::NoDropShadowWindowHint),
         items_(items),
-        zoom_(qMax(1, zoomPercent) / 100.0),
+        zoom_(qMax(qreal(1), zoomPercent) / 100.0),
         metrics_(chromeMenuMetrics(zoom_)),
         look_(look) {
     setAttribute(Qt::WA_TranslucentBackground);
@@ -250,7 +250,7 @@ class ChromeMenuWindow : public QWidget {
 }  // namespace
 
 int execChromeMenu(QWidget* owner, const QVector<ChromeMenuItem>& items,
-                   const QRect& anchorGlobal, PopupAnchor anchor, int zoomPercent,
+                   const QRect& anchorGlobal, PopupAnchor anchor, qreal zoomPercent,
                    const ChromeTokens& look) {
   if (!owner || items.isEmpty()) return kChromeMenuNone;
   ChromeMenuWindow menu(owner, items, zoomPercent, look);
